grub-efi_%.bbappend: Add internal media grub.cfg_internal to deploy directory
We need the bootloader config file for when booting from internal media as now
we do not change the bootloader specific files when the flasher runs. What the
flasher does now is simply copy over to the internal media boot partition this
file we deploy here. This way, we keep the flasher free from any board specific
changes.

Signed-off-by: Florin Sarbu <florin@resin.io>