resin-image-flasher.bbappend: Specify addition of the internal media bootloader config file to the resin-image-flasher boot partition.
We need this bootloader config file available to the flasher when it provisions
a device so that on next reboot from internal media it can pick-up the correct boot
options as needed.

We also handle from here adding the grub.cfg file for external media boot to the correct
location expected by EFI grub when booting from the USB key. Also, we renamed this grub.cfg
for external media booting (USB key) to grub.cfg_external.

Signed-off-by: Florin Sarbu <florin@resin.io>