grub: Add Automated Rollback support to grub.cfg
Automated Rollback support in meta-resin is used to recover the device
in case a HUP fails. The rollback services rely on resin_root_part
and upgrade_available variables to be present in the grub.cfg file.

Their usage is similar to the u-boot use of these variables.

Bootcount is incremented in grubenv before loading the kernel.
If a boot fails, grub will try to load the kernel from the previous
rootfs.

If boot succeeds, the rollback systemd services will check and
update upgrade_available and resin_root_part in the grub.cfg file

The strange logic to store bootcount is due to limited scripting
options in grub environment.

Changelog-entry: Add Automated Rollback support to grub.cfg
Signed-off-by: Zubair Lutfullah Kakakhel <zubair@resin.io>