resin-image: Install the legacy grub.cfg in boot files
Any changes to the grub.cfg file are not updated in the legacy
grub. Add the grub.cfg file in the bootfiles so that grub.cfg
is updated during a HUP.

The HUP hook needs to be updated to remove the redundant grub.cfg
file.

Changelog-entry: Install legacy grub.cfg file in boot files
Signed-off-by: Zubair Lutfullah Kakakhel <zubair@resin.io>