grub: Temporarily disable rollbacks
At this point we need to figure out more scenarios that would trigger
a reboot before healthcheck actually runs (such as supervisor triggered
reboots).
Such an early reboot is not an indication of a hostOS update failure and
so we do not wish for rollbacks to take place in these scenarios.

Until all corner cases are discussed and handled in the OS, we disable
rollback functionality.

Changelog-entry: Temporarily disable rollback functionality
Signed-off-by: Florin Sarbu <florin@balena.io>