resin-init-board: Add script to ensure correct future EFI boot
It's been observed on EFI machines that at the first boot after
provisioning there is some bootstrap code in the first sector of
the boot media (even though during provisioning we make sure we
remove this bootstrap code). In such a case (only on some machines),
the next time we reboot the machine, it will not recognize the media
as a valid EFI boot media and the boot will fail.

Changelog-entry: Add script to ensure correct future EFI boot
Signed-off-by: Florin Sarbu <florin@balena.io>