linux-yocto_%.bbappend: Specify deps for config tasks
We switched to kernel 4.18.14. For newer kernels (4.17+), we need
to specify additional dependencies for the configme task.

This change can be removed when we update poky to a newer version
that includes 81e8a52e8e40e47c34f900db5d73e69ffc25f5d0.

This change fixes the error we get on 4.18.14 kernel when our do
kernel_resin_reconfigure task will recreate the .config with the
kernel config options from the development machine's kernel config.

Signed-off-by: Florin Sarbu <florin@balena.io>