kernel-modules-headers: Fix compile issue on kernel 5.0.3
We need to make sure the cross fixdep is used when compiling
the target objtool binary on kernel 5.0.3 which is the kernel
version we have since we updated to Poky Warrior.
This commit fixes the following error:

 /bin/sh: 1: /yocto/resin-board/build/tmp/work/
genericx86_64-poky-linux/kernel-modules-headers/
1.0-r0/kernel_modules_headers/tools/objtool//fixdep: not found

The above error is encountered because the compile stage is using
the target fixdep binary which has the target intepreter for
such a case when running it on the build machine.
(this commit is to be removed from here when it gets merged to
meta-balena and we incorporate thta meta-balena release)

Signed-off-by: Florin Sarbu <florin@balena.io>