grub-conf: Include in cmdline partitions UUID when available
After first boot, the OS will generate the UUID of every partition and
store their values in a boot partition file. Source this file and pass
these UUIDs, when available, as part of cmdline so that the OS can use
them for partition mounts.

Changelog-entry: Pass partitions UUID in cmdline when available
Signed-off-by: Andrei Gherzan <andrei@balena.io>