grub-conf: Select root device from the same drive as boot media
This commit parses all partitions in the boot media looking for the
corresponding resin-root label and passes its UUID to the kernel command
line.

This guarantees the system will always boot from the same drive as grub
is booting from.

Note that the maximum number of partitions is will search for is 10.

Changelog-entry: Grub selects root device from the same drive as boot device
Signed-off-by: Alex Gonzalez <alexg@balena.io>