surface-go: restore vendor firmware for qca6174
It seems the Surface Go ships a variant of the QCA6174 wireless adapter
that doesn't work with the same firmware used upstream for this chipset
everywhere else. There's a vendor supplied firmware that works on the
Surface Go, but this device specific vendor firmware is incompatible
with other devices.

https://bugs.launchpad.net/ubuntu/+source/linux-firmware/+bug/1804028

Install this broken firmware only on the Surface Go to fix wireless, and
use upstream firmware in all other cases.

Changelog-entry: surface-go: restore vendor firmware for qca6174
Signed-off-by: Joseph Kogut <joseph@balena.io>