balena-image-initramfs: Increase max size to 64M on genericx86-64-ext
This is necessary to accomodate cryptsetup. The former 32MB limit comes from
a limitation for devices that need to fit the initramdisk into a 40MB boot
partition but this is not the case for genericx86-64-ext where the initramdisk
comes bundled together with the kernel in the root partition.

Change-type: patch
Signed-off-by: Michal Toman <michalt@balena.io>