genericx86-64-ext: Make VMD driver built-in
This is a driver that adds a PCI bridge that provides access
to (mostly storage) devices "located in a different PCI domain".

Without the driver/bridge, such devices are invisible to the kernel,
which may trick the migrator into thinking there are no valid flashing
targets and start the in-memory migration.

This patch makes the driver built-in, which means the devices will
get initialized before the migrator checks for potential targets.

Changelog-entry: Initialize VMD devices before the migrator runs
Signed-off-by: Michal Toman <michalt@balena.io>