balena-image-initramfs: Remove migrate module for Intel NUC
because this device-type is not a target for migration and
is running out of space for the rust update.

We do keep the module for the genericx86-64-ext because it
extends the genericx86-64 yocto machine and has not run out
of space.

Changelog-entry: balena-image-initramfs: Remove migrate module for Intel NUC
Signed-off-by: Alexandru Costache <alexandru@balena.io>